Add new component with configurable image
The new AppBanner component also has configurable tertiary text. The old
DigitalSubscriberAppBanner component has been refactored to use the more
generic AppBanner.

The image must match one of an allow-list of URL prefixes which we've
specified (currently https://media.guim.co.uk/ and https://i.gium.co.uk)
to prevent any old arbitrary images from being used.

The plan is to add a Grid picker to Storybook to make it easier to add
an image.
