+++ .storybook/main.js
@@ -1,6 +1,9 @@
-const path = require("path");
+const path = require('path');
 
 module.exports = {
-  stories: ["../src/**/*.stories.tsx"],
-  addons: ["@storybook/addon-knobs", "@storybook/addon-viewport","./gu-preview/register.js"],
+    stories:
+        process.env.NODE_ENV === 'development'
+            ? ['../src/**/*.stories.tsx']
+            : ['../src/!(AppBanner)/*.stories.tsx'],
+    addons: ['@storybook/addon-knobs', '@storybook/addon-viewport', './gu-preview/register.js'],
 };
